pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.backends._legacy:_Backend"

[project]
name = "clarity-cx"
version = "1.0.0"
description = "AI-Powered Call Center Intelligence Platform"
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.11"
authors = [
    {name = "Clarity CX Team"}
]

dependencies = [
    "streamlit>=1.39.0",
    "fastapi>=0.115.0",
    "uvicorn>=0.32.0",
    "langgraph>=0.2.0",
    "langchain>=0.3.0",
    "langchain-openai>=0.2.0",
    "langchain-anthropic>=0.2.0",
    "langchain-google-genai>=2.0.0",
    "langchain-community>=0.3.0",
    "openai>=1.50.0",
    "anthropic>=0.37.0",
    "google-generativeai>=0.8.0",
    "pydantic>=2.9.0",
    "chromadb>=0.5.0",
    "plotly>=5.24.0",
    "python-dotenv>=1.0.0",
    "python-multipart>=0.0.9",
    "httpx>=0.27.0",
    "aiofiles>=24.1.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.3.0",
    "pytest-asyncio>=0.24.0",
    "pytest-cov>=5.0.0",
    "deepeval>=1.0.0",
    "ruff>=0.7.0",
]
observability = [
    "arize-phoenix>=8.0.0",
    "opentelemetry-api>=1.20.0",
    "opentelemetry-sdk>=1.20.0",
    "opentelemetry-exporter-otlp>=1.20.0",
    "openinference-instrumentation-langchain>=0.1.0",
]

[tool.setuptools.packages.find]
where = ["."]
include = ["src*"]

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"

[tool.ruff]
line-length = 100
target-version = "py311"
